+++ backend/tests/db-data.setup.ts
@@ -19,7 +19,9 @@ export class DBTestData {
             users: 'users',
             tickets: 'tickets',
             rateLimits: 'rate_limits',
-            demoLimits: 'demo_limits'
+            demoLimits: 'demo_limits',
+            feedback: 'feedback_entries',
+            feedbackRating: 'feedback_ratings'
         };
     }
 
@@ -47,27 +49,36 @@ export class DBTestData {
     getClientsInsertSql(): BaseQuery {
         const sql = `INSERT INTO ${this.tableRecords['clients']}
         (client_id, name, api_key_hash, status, flag, last_use, last_modified, created_on)
-        VALUES ($1, $2, $3, $4, $5, $6, $7, $8);
+        VALUES ($1, $2, $3, $4, $5, $6, $7, $8), ($9, $10, $11, $12, $13, $14, $15, $16);
         `;
-        const values = [mockId.clients.valid[0], 'TESTCLIENT', secrets.TEST_APIKEY_HASH, ApiKeyStatus.ACTIVE, null, '2025-01-01T14:00:02.000Z', '2025-01-01T14:00:02.000Z', '2025-01-01T14:00:02.000Z'];
+        const values = [
+            mockId.clients.valid[0], 'TESTCLIENT', secrets.TEST_APIKEY_HASH, ApiKeyStatus.ACTIVE, null, '2025-01-01T14:00:02.000Z', '2025-01-01T14:00:02.000Z', '2025-01-01T14:00:02.000Z',
+            mockId.clients.valid[1], 'ANOTHER-TESTCLIENT', secrets.TEST_APIKEY_HASH, ApiKeyStatus.ACTIVE, null, '2025-01-01T14:00:02.000Z', '2025-01-01T14:00:02.000Z', '2025-01-01T14:00:02.000Z'
+        ];
         return { sql: sql, values: values };
     }
 
     getUsersInsertSql(): BaseQuery {
         const sql = `INSERT INTO ${this.tableRecords['users']}
         (user_id, email, status, flag, last_modified, created_on)
-        VALUES ($1, $2, $3, $4, $5, $6);
+        VALUES ($1, $2, $3, $4, $5, $6), ($7, $8, $9, $10, $11, $12);
         `;
-        const values = [mockId.users.valid[0], 'max.mustermann@yqni13.com', UserStatus.ACTIVE, null, '2025-01-01T14:00:03.000Z', '2025-01-01T14:00:03.000Z'];
+        const values = [
+            mockId.users.valid[0], 'max.mustermann@yqni13.com', UserStatus.ACTIVE, null, '2025-01-01T14:00:03.000Z', '2025-01-01T14:00:03.000Z',
+            mockId.users.valid[1], 'aurora.otsuki@yqni13.com', UserStatus.ACTIVE, null, '2026-01-01T14:00:03.000Z', '2026-01-01T14:00:03.000Z'
+        ];
         return { sql: sql, values: values };
     }
 
     getTicketsInsertSql(): BaseQuery {
         const sql = `INSERT INTO ${this.tableRecords['tickets']}
         (ticket_id, client_id, user_id, status, option, title, message, resource_paths, flag, info_browser, info_os, info_device, last_modified, created_on)
-        VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14);
+        VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14), ($15, $16, $17, $18, $19, $20, $21, $22, $23, $24, $25, $26, $27, $28);
         `;
-        const values = [mockId.tickets.valid[0], mockId.clients.valid[0], mockId.users.valid[0], TicketStatus.ISSUED, TicketOption.SUPPORT, 'test-title', 'test-message', ['test/path/num0', 'test/path/num1'], null, 'Brave 1.87.190 (Official Build) (64-Bit)', 'Windows 11', DeviceOption.COMPUTER, '2025-01-01T14:00:04.000Z', '2025-01-01T14:00:04.000Z'];
+        const values = [
+            mockId.tickets.valid[0], mockId.clients.valid[0], mockId.users.valid[0], TicketStatus.ISSUED, TicketOption.SUPPORT, 'test-title', 'test-message', ['test/path/num0', 'test/path/num1'], null, 'Brave 1.87.190 (Official Build) (64-Bit)', 'Windows 11', DeviceOption.COMPUTER, '2025-01-01T14:00:04.000Z', '2025-01-01T14:00:04.000Z',
+            mockId.tickets.valid[1], mockId.clients.valid[0], mockId.users.valid[0], TicketStatus.ISSUED, TicketOption.SUPPORT, 'test-title', 'test-message-without-resource_paths', undefined, null, 'Brave 1.87.190 (Official Build) (64-Bit)', 'Android 15', DeviceOption.MOBILE, '2025-01-01T14:00:07.000Z', '2025-01-01T14:00:07.000Z'
+        ];
         return { sql: sql, values: values };
     }
 
@@ -89,12 +100,21 @@ export class DBTestData {
         return { sql: sql, values: values }; 
     }
 
-    getTicketsWithoutPathsInsertSql(): BaseQuery {
-        const sql = `INSERT INTO ${this.tableRecords['tickets']}
-        (ticket_id, client_id, user_id, status, option, title, message, resource_paths, flag, info_browser, info_os, info_device, last_modified, created_on)
-        VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14);
+    getFeedbackInsertSql(): BaseQuery {
+        const sql = `INSERT INTO ${this.tableRecords['feedback']}
+        (client_id, user_id, rating, term_accepted, message, reviewed_on, last_modified, created_on)
+        VALUES ($1, $2, $3, $4, $5, $6, $7, $8);
         `;
-        const values = [mockId.tickets.valid[1], mockId.clients.valid[0], mockId.users.valid[0], TicketStatus.ISSUED, TicketOption.SUPPORT, 'test-title', 'test-message-without-resource_paths', undefined, null, 'Brave 1.87.190 (Official Build) (64-Bit)', 'Android 15', DeviceOption.MOBILE, '2025-01-01T14:00:07.000Z', '2025-01-01T14:00:07.000Z'];
+        const values = [mockId.clients.valid[0], mockId.users.valid[0], 5, true, 'test-feedback-message', undefined, '2025-01-01T14:00:08.000Z', '2025-01-01T14:00:08.000Z'];
+        return { sql: sql, values: values };
+    }
+
+    getFeedbackRatingInsertSql(): BaseQuery {
+        const sql = `INSERT INTO ${this.tableRecords['feedbackRating']}
+        (client_id, count, rating_sum, last_modified, created_on)
+        VALUES ($1, $2, $3, $4, $5);
+        `;
+        const values = [mockId.clients.valid[0], 16, 67, '2025-01-01T14:00:09.000Z', '2025-01-01T14:00:09.000Z'];
         return { sql: sql, values: values };
     }
 }